# [common][linux]: improve readlines performance.