… internal/common/testdata/linux/issue_1514/stat