@atoulme 2023-08-28
I see different results on my machine with 80 cores. Here they are:
```
goos: linux
goarch: arm64
pkg: github.com/shirou/gopsutil/v3/process
             │ old1515.txt │             new1515.txt             │
             │   sec/op    │   sec/op     vs base                │
Processes-80   249.4m ± 2%   297.3m ± 2%  +19.19% (p=0.000 n=10)
```
This is with go 1.21.0.