@shirou 2023-08-29
I have tested with AWS c5a.16xlarge (CPU=64).  and I got no total performance improvement. hmm...

```
goos: linux
goarch: amd64
pkg: github.com/shirou/gopsutil/v3/process
cpu: AMD EPYC 7R32
             │  base.txt   │             after.txt              │
             │   sec/op    │   sec/op     vs base               │
Processes-64   37.55m ± 1%   38.32m ± 0%  +2.06% (p=0.002 n=10)

             │   base.txt   │               after.txt               │
             │     B/op     │     B/op       vs base                │
Processes-64   9.266Mi ± 0%   14.602Mi ± 0%  +57.59% (p=0.000 n=10)

             │  base.txt   │              after.txt              │
             │  allocs/op  │  allocs/op   vs base                │
Processes-64   64.68k ± 0%   25.49k ± 0%  -60.59% (p=0.000 n=10)
```